requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn>=0.29.0
httpx>=0.27.0
pydantic-settings>=2.2.1

# LangGraph & LLM
langgraph>=1.1.6
langchain-core>=0.3.0
langchain-google-vertexai>=2.0.0

# Database
asyncpg>=0.29.0
SQLAlchemy>=2.0.29
alembic>=1.13.1

# Testing & Utils
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.5